:truck: create table folder
