+++ js/table/table.js
@@ -1,7 +1,10 @@
-import { setUpDeleteAllModal } from "./shots/delete-all-modal.js";
-import { getDefaultDetails } from "./details/config-details.js";
+import { setUpDeleteAllModal } from "../shots/delete-all-modal.js";
+import { getDefaultDetails } from "../details/config-details.js";
+import { getRows } from "./table-functions.js";
 
 function setUpTable() {
+    sessionStorage.setItem("rows", JSON.stringify([]));
+
     d3.select("#shot-table")
         .append("thead")
         .append("tr");
@@ -12,7 +15,29 @@ function setUpTable() {
         .append("tbody")
         .attr("id", "shot-table-body");
 
+    d3.select("#shot-table")
+        .append("tfoot")
+        .append("tr");
+
+    createFooterRow();
+
     setUpDeleteAllModal("#delete-all-modal");
+}
+
+function createFooterRow() {
+    var footerRow = d3
+        .select("#shot-table")
+        .select("tfoot")
+        .select("tr")
+        .attr("class", "small-text");
+
+    var t = footerRow.append("td").attr("colspan", 9); // TODO: make dynamic
+
+    console.log(getRows());
+
+    let totalRowCount = getRows().length;
+
+    t.text(`Displaying ${1} - ${10} of ${totalRowCount} rows.`);
 }
 
 function createHeaderRow(details) {
@@ -46,34 +71,4 @@ function createHeaderRow(details) {
         });
 }
 
-function clearTable() {
-    sessionStorage.setItem("rows", JSON.stringify([]));
-
-    d3.select("#shot-table-body")
-        .selectAll("tr")
-        .remove();
-
-    var dots = d3.select("#hockey-rink-svg").select("#dots");
-
-    dots.select("#normal")
-        .selectAll("*")
-        .remove();
-    dots.select("#selected")
-        .selectAll("*")
-        .remove();
-}
-
-function getHeaderRow() {
-    var l = [];
-    d3.select("#shot-table")
-        .select("thead")
-        .selectAll("th")
-        .each(function() {
-            let dataId = d3.select(this).attr("data-id");
-            let dataType = d3.select(this).attr("data-type");
-            l.push({ id: dataId, type: dataType });
-        });
-    return l;
-}
-
-export { setUpTable, clearTable, createHeaderRow, getHeaderRow };
+export { setUpTable, createHeaderRow };
